scripts: collect examples
